🩹 Move to NumberChip component Chip folder
NumberChip 컴포넌트 Chip 폴더로 이동
