+++ app/(dashboard)/dashboard/[id]/page.tsx
@@ -1,9 +1,7 @@
+import NumberChip from '@/components/Chip/NumberChip';
 import AddButton from '@/components/dashboard/AddButton/AddButton';
 import DashboardList from '@/components/dashboard/DashboardList/DashboardList';
-import NumberChip from '@/components/dashboard/DashboardTitle/NumberChip'; // 임시 컴포넌트
 import TEMP from '@/public/image/desktop.jpg'; //임시 이미지
-
-// interface폳더로 옮길 dashboardProps
 
 const Dashboard = () => {
   return (
